requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
drf-spectacular==0.27.0
django-cors-headers==4.3.1
